1. Add HSD backprojection as another option Now you can choose from 3 options: "H HS HSD" to do the backprojection.
The HSD model is slower(6ms for one object), it still can be improved.
If the objects are in different color ,choose HS rathar than HSD

2.
Modify the YAML file(multiple_objects_detection.yaml) and add some
comments

3.
Delete the "multiple_objects_detectionConfig.h" cause it can be generate
by the CFG file .

4.
Set a varable camera_name to kinect2_head to const, we can use launch
file to remap the kinect2_head based topic with the Camera_name in the
detection launch file.

5.
After you select the objects, the window which show the detections will
be destroyed. With this you can draw the bounding box easier in the
Master computer cause ther won't be a lot of windsows displaying the
images from every camera.

If you still want to check the detections on the image, set the varible
"show_2D_track" in the "multiple_objects_detection.yaml" to true.